+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/board/EditBoardListener.java
@@ -1,0 +1,9 @@
+package it.niedermann.nextcloud.deck.ui.board;
+
+import it.niedermann.nextcloud.deck.model.full.FullBoard;
+
+public interface EditBoardListener {
+    void onUpdateBoard(FullBoard fullBoard);
+
+    void onCreateBoard(String title, String color);
+}
