+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/board/accesscontrol/AccessControlChangedListener.java
@@ -1,0 +1,9 @@
+package it.niedermann.nextcloud.deck.ui.board.accesscontrol;
+
+import it.niedermann.nextcloud.deck.model.AccessControl;
+
+public interface AccessControlChangedListener {
+        void updateAccessControl(AccessControl accessControl);
+
+        void deleteAccessControl(AccessControl ac);
+    }
