+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/card/CardAdapter.java
@@ -46,8 +46,8 @@ import it.niedermann.nextcloud.deck.model.full.FullStack;
 import it.niedermann.nextcloud.deck.persistence.sync.SyncManager;
 import it.niedermann.nextcloud.deck.persistence.sync.adapters.db.util.LiveDataHelper;
 import it.niedermann.nextcloud.deck.ui.EditActivity;
-import it.niedermann.nextcloud.deck.ui.helper.dnd.DragAndDropAdapter;
-import it.niedermann.nextcloud.deck.ui.helper.dnd.DraggedItemLocalState;
+import it.niedermann.nextcloud.deck.ui.dnd.DragAndDropAdapter;
+import it.niedermann.nextcloud.deck.ui.dnd.DraggedItemLocalState;
 import it.niedermann.nextcloud.deck.util.ColorUtil;
 import it.niedermann.nextcloud.deck.util.DateUtil;
 import it.niedermann.nextcloud.deck.util.DimensionUtil;
