+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/card/attachments/AttachmentClickedListener.java
@@ -1,0 +1,5 @@
+package it.niedermann.nextcloud.deck.ui.card.attachments;
+
+public interface AttachmentClickedListener {
+        void onAttachmentClicked(int position);
+    }
