+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/card/attachments/AttachmentDeletedListener.java
@@ -1,0 +1,7 @@
+package it.niedermann.nextcloud.deck.ui.card.attachments;
+
+import it.niedermann.nextcloud.deck.model.Attachment;
+
+public interface AttachmentDeletedListener {
+        void onAttachmentDeleted(Attachment attachment);
+    }
