+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/card/attachments/NewCardAttachmentHandler.java
@@ -1,0 +1,9 @@
+package it.niedermann.nextcloud.deck.ui.card.attachments;
+
+import it.niedermann.nextcloud.deck.model.Attachment;
+
+public interface NewCardAttachmentHandler {
+    void attachmentAdded(Attachment attachment);
+
+    void attachmentRemoved(Attachment attachment);
+}
