+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/dnd/DragAndDropModel.java
@@ -1,4 +1,4 @@
-package it.niedermann.nextcloud.deck.ui.helper.dnd;
+package it.niedermann.nextcloud.deck.ui.dnd;
 
 import androidx.annotation.NonNull;
