+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/dnd/DragAndDropTab.java
@@ -1,4 +1,4 @@
-package it.niedermann.nextcloud.deck.ui.helper.dnd;
+package it.niedermann.nextcloud.deck.ui.dnd;
 
 import androidx.recyclerview.widget.RecyclerView;
