+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/dnd/DraggedItemLocalState.java
@@ -1,4 +1,4 @@
-package it.niedermann.nextcloud.deck.ui.helper.dnd;
+package it.niedermann.nextcloud.deck.ui.dnd;
 
 import android.view.View;
