+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/dnd/ItemMovedByDragListener.java
@@ -1,4 +1,4 @@
-package it.niedermann.nextcloud.deck.ui.helper.dnd;
+package it.niedermann.nextcloud.deck.ui.dnd;
 
 public interface ItemMovedByDragListener<ItemModel> {
     void onItemMoved(ItemModel movedItem, long tabId, int position);
