+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/dnd/ScrollHelper.java
@@ -1,4 +1,4 @@
-package it.niedermann.nextcloud.deck.ui.helper.dnd;
+package it.niedermann.nextcloud.deck.ui.dnd;
 
 import android.os.Handler;
