+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/stack/EditStackListener.java
@@ -1,0 +1,7 @@
+package it.niedermann.nextcloud.deck.ui.stack;
+
+public interface EditStackListener {
+    void onCreateStack(String title);
+
+    void onUpdateStack(long stackId, String title);
+}
