+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/stack/OnScrollListener.java
@@ -1,0 +1,7 @@
+package it.niedermann.nextcloud.deck.ui.stack;
+
+public interface OnScrollListener {
+    void onScrollUp();
+
+    void onScrollDown();
+}
